requirements.txt
streamlit
google-generativeai
python-dotenv
transformers
python-chess
